// src/types.tsx
export type FormField = {
  type:
    | 'text'
    | 'file'
    | 'files'
    | 'radio'
    | 'dropdown'
    | 'location'
    | 'subdomain'
    | 'password'
    | 'phone'
    | 'date'
    | 'currency'
    | 'textarea'
    | 'image'
    | 'images'
    | 'duration'
    | 'schedule'
    | 'service'
    | 'selectPayment'
    | 'shipping'
    | 'checkout';
  id: string;
  label: string;
  name: string;
  autocomplete?: string;
  onButtonClick?: () => void;
  min?: number;
  max?: number;
  pattern?: RegExp;
  required?: boolean;
  disabled?: boolean;
  icon?: string;
  // Dropdown props
  opts?: DropDownOpt[];
  multi?: boolean;
  errorMessage?: string;
  formatLabel?: (obj: DropDownOpt) => ReactNode;
  /**
   * RegExpression to format data onSubmit
   * @example value.replace(formatter, "")
   */
  formatter?: RegExp;
};

export type Form = {
  fields: FormField[];
  onSubmit: () => any;
};

export type FormInputElement = HTMLInputElement & { files: File[] };